the last primitive generators

+++ QuickMGenerate/DecimalGen.cs
@@ -9,7 +9,7 @@ namespace QuickMGenerate
 			return Decimal(1, 100);
 		}
 
-		public static Generator<State, decimal> Decimal(int min, int max)
+		public static Generator<State, decimal> Decimal(decimal min, decimal max)
 		{
 			return s => new Result<State, decimal>(((decimal)s.Random.NextDouble() * (max - min)) + min, s);
 		}
